requirements.txt


flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0
numpy==1.26.0
pandas==2.1.0
scipy==1.11.0
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.18.0
networkx==3.2.0
pyyaml==6.0.1
transformers==4.35.0
tokenizers==0.15.0
onnx==1.15.0
onnxruntime==1.16.0
scikit-learn==1.3.0
tqdm==4.66.0
rich==13.6.0
pydantic==2.4.0
omegaconf==2.3.0
psutil==5.9.0